@@ -2,7 +2,7 @@
   "name": "accutheme",
   "version": "1.0.0",
   "description": "Starter theme for DNN Platform websites",
-  "main": "gulpfile.js",
+  "exports": "./gulpfile.js",
   "scripts": {
     "build": "cross-env NODE_ENV=production gulp build",
     "clean": "gulp clean",
@@ -12,6 +12,10 @@
     "lint:styles": "stylelint --syntax scss",
     "postinstall": "gulp build",
     "package": "webpack --config webpack.installer.config.js"
+  },
+  "type": "module",
+  "engines": {
+    "node": "^12.20.0 || ^14.13.1 || >=16.0.0"
   },
   "private": true,
   "repository": {
@@ -23,23 +27,21 @@
   "dependencies": {
     "bootstrap": "^4.6.0",
     "bootstrap-icons": "^1.5.0",
-    "core-js": "^3.16.4",
-    "file-loader": "^6.2.0",
+    "core-js": "^3.17.2",
     "flickity": "^2.2.2",
     "jquery": "^3.6.0",
-    "pngquant-bin": "^3.1.1",
     "popper.js": "^1.16.1",
     "srraf": "^3.2.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.15.0",
-    "@babel/eslint-parser": "^7.15.0",
+    "@babel/core": "^7.15.5",
+    "@babel/eslint-parser": "^7.15.4",
     "@babel/eslint-plugin": "^7.14.5",
     "@babel/plugin-proposal-class-properties": "^7.14.5",
     "@babel/plugin-transform-runtime": "^7.15.0",
-    "@babel/preset-env": "^7.15.0",
-    "autoprefixer": "^10.3.3",
-    "acorn": "^8.4.1",
+    "@babel/preset-env": "^7.15.4",
+    "acorn": "^8.5.0",
+    "autoprefixer": "^10.3.4",
     "babel-loader": "^8.2.2",
     "cross-env": "^7.0.3",
     "cssnano": "^5.0.8",
@@ -48,19 +50,23 @@
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-formatter-pretty": "^4.1.0",
-    "eslint-plugin-prettier": "^3.4.1",
+    "eslint-plugin-prettier": "^4.0.0",
     "eslint-webpack-plugin": "^3.0.1",
+    "esm": "^3.2.25",
+    "file-loader": "^6.2.0",
     "filemanager-webpack-plugin": "^6.1.6",
     "gulp": "^4.0.2",
-    "gulp-imagemin": "^7.1.0",
+    "gulp-cli": "^2.3.0",
+    "gulp-imagemin": "^8.0.0",
     "gulp-postcss": "^9.0.1",
     "gulp-real-favicon": "^0.3.2",
     "gulp-replace": "^1.1.3",
-    "gulp-sass": "^4.1.0",
+    "gulp-sass": "^5.0.0",
     "gulp-sourcemaps": "^3.0.0",
     "gulp-stylelint": "^13.0.0",
     "husky": "^7.0.2",
-    "imagemin-pngquant": "^5.0.1",
+    "imagemin-pngquant": "^9.0.2",
+    "imagemin-svgo": "^9.0.0",
     "lint-staged": "^11.1.2",
     "node-sass": "^6.0.1",
     "postcss": "^8.3.6",
@@ -68,18 +74,15 @@
     "prettier": "^2.3.2",
     "stylelint": "^13.13.1",
     "stylelint-config-twbs-bootstrap": "^2.2.3",
-    "stylelint-formatter-pretty": "^2.1.1",
-    "terser-webpack-plugin": "^5.1.4",
-    "webpack": "^5.51.1",
+    "stylelint-formatter-pretty": "^3.0.0",
+    "terser-webpack-plugin": "^5.2.3",
+    "webpack": "^5.52.0",
     "webpack-cli": "^4.8.0"
   },
   "husky": {
     "hooks": {
       "pre-commit": "lint-staged"
     }
-  },
-  "resolutions": {
-    "image-webpack-loader/imagemin-pngquant": "5.0.1"
   },
   "lint-staged": {
     "*.js": [
